swarms/researcher_swarm/requirements.txt
#swarmzero==0.0.1
-e ../../../swarmzero
firecrawl-py==1.4.0
serpapi==0.1.5
google-search-results==2.4.2
msal==1.31.0
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
aiosqlite==0.20.0
python-dotenv==1.0.1
requests==2.32.3
fpdf2==2.8.1
python-docx==1.1.2
